Update README.md
It is a one to one and multi to one chat system. It is built by using concepts such as Socket Programming, Multhithreading, OOPs concepts. This chat system can handle any number of clients. All the clients which are logged in to system are shown in online list and any client can chat with any other online client.